+++ src/radical/pilot/__init__.py
@@ -27,7 +27,7 @@ from .compute_pilot             import ComputePilot
 from .compute_pilot_description import ComputePilotDescription
 
 from .resource_config           import ResourceConfig
-from .staging_directives        import COPY, LINK, MOVE, TRANSFER, 
+from .staging_directives        import COPY, LINK, MOVE, TRANSFER
 from .staging_directives        import SKIP_FAILED, CREATE_PARENTS
 
 from .utils                     import version, version_detail, version_branch
