+++ src/radical/pilot/agent/__init__.py
@@ -1,3 +1,8 @@
 
-from .rm import RM
+__copyright__ = "Copyright 2013-2014, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .rm        import RM
+from .scheduler import Scheduler
