+++ src/radical/pilot/agent/radical-pilot-agent-multicore.py
@@ -218,12 +218,6 @@ LAUNCH_METHOD_RUNJOB        = 'RUNJOB'
 LAUNCH_METHOD_SSH           = 'SSH'
 LAUNCH_METHOD_YARN          = 'YARN'
 
-# 'enum' for pilot's unit scheduler types
-SCHEDULER_NAME_CONTINUOUS   = "CONTINUOUS"
-SCHEDULER_NAME_SCATTERED    = "SCATTERED"
-SCHEDULER_NAME_TORUS        = "TORUS"
-SCHEDULER_NAME_YARN         = "YARN"
-
 # 'enum' for pilot's unit spawner types
 SPAWNER_NAME_POPEN          = "POPEN"
 SPAWNER_NAME_SHELL          = "SHELL"
@@ -254,10 +248,6 @@ TRANSFER = 'Transfer' # saga remote transfer
 OK       = 'OK'
 FAIL     = 'FAIL'
 RETRY    = 'RETRY'
-
-# two-state for slot occupation.
-FREE     = 'Free'
-BUSY     = 'Busy'
 
 # ----------------------------------------------------------------------------------
 #
@@ -403,1011 +393,6 @@ def pilot_DONE(mongo_p=None, pilot_uid=None, logger=None, msg=None):
             logger.error("cannot log cancel state in database!")
 
         print "cannot log cancel state in database!"
-
-
-
-# ==============================================================================
-#
-# Schedulers
-#
-# ==============================================================================
-#
-class AgentSchedulingComponent(rpu.Component):
-
-    # FIXME: clarify what can be overloaded by Scheduler classes
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        rpu.Component.__init__(self, 'AgentSchedulingComponent', cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def initialize_child(self):
-
-      # self.declare_input (rp.AGENT_SCHEDULING_PENDING, rp.AGENT_SCHEDULING_QUEUE)
-      # self.declare_worker(rp.AGENT_SCHEDULING_PENDING, self.work)
-
-        self.declare_input (rp.ALLOCATING_PENDING, rp.AGENT_SCHEDULING_QUEUE)
-        self.declare_worker(rp.ALLOCATING_PENDING, self.work)
-
-        self.declare_output(rp.EXECUTING_PENDING,  rp.AGENT_EXECUTING_QUEUE)
-
-        # we need unschedule updates to learn about units which free their
-        # allocated cores.  Those updates need to be issued after execution, ie.
-        # by the AgentExecutionComponent.
-        self.declare_publisher ('state',      rp.AGENT_STATE_PUBSUB)
-        self.declare_subscriber('unschedule', rp.AGENT_UNSCHEDULE_PUBSUB, self.unschedule_cb)
-
-        # we create a pubsub pair for reschedule trigger
-        self.declare_publisher ('reschedule', rp.AGENT_RESCHEDULE_PUBSUB)
-        self.declare_subscriber('reschedule', rp.AGENT_RESCHEDULE_PUBSUB, self.reschedule_cb)
-
-        # all components use the command channel for control messages
-        self.declare_publisher ('command', rp.AGENT_COMMAND_PUBSUB)
-
-        # we declare a clone and a drop callback, so that cores can be assigned
-        # to clones, and can also be freed again.
-        self.declare_clone_cb(self.clone_cb)
-        self.declare_drop_cb (self.drop_cb)
-
-        # when cloning, we fake scheduling via round robin over all cores.
-        # These indexes keeps track of the last used core.
-        self._clone_slot_idx = 0
-        self._clone_core_idx = 0
-
-        # The scheduler needs the LRMS information which have been collected
-        # during agent startup.  We dig them out of the config at this point.
-        self._pilot_id = self._cfg['pilot_id']
-        self._lrms_lm_info        = self._cfg['lrms_info']['lm_info']
-        self._lrms_node_list      = self._cfg['lrms_info']['node_list']
-        self._lrms_cores_per_node = self._cfg['lrms_info']['cores_per_node']
-        # FIXME: this information is insufficient for the torus scheduler!
-
-        self._wait_pool = list()            # set of units which wait for the resource
-        self._wait_lock = threading.RLock() # look on the above set
-        self._slot_lock = threading.RLock() # look for slot allocation/deallocation
-
-        # configure the scheduler instance
-        self._configure()
-
-        # communicate successful startup
-        self.publish('command', {'cmd' : 'alive',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    def finalize_child(self):
-
-        # communicate finalization
-        self.publish('command', {'cmd' : 'final',
-                                 'arg' : self.cname})
-
-
-    # --------------------------------------------------------------------------
-    #
-    # This class-method creates the appropriate sub-class for the Scheduler.
-    #
-    @classmethod
-    def create(cls, cfg):
-
-        # Make sure that we are the base-class!
-        if cls != AgentSchedulingComponent:
-            raise TypeError("Scheduler Factory only available to base class!")
-
-        name = cfg['scheduler']
-
-        try:
-            impl = {
-                SCHEDULER_NAME_CONTINUOUS : SchedulerContinuous,
-                SCHEDULER_NAME_SCATTERED  : SchedulerScattered,
-                SCHEDULER_NAME_TORUS      : SchedulerTorus,
-                SCHEDULER_NAME_YARN       : SchedulerYarn
-            }[name]
-
-            impl = impl(cfg)
-            return impl
-
-        except KeyError:
-            raise ValueError("Scheduler '%s' unknown or defunct" % name)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        raise NotImplementedError("_configure() not implemented for Scheduler '%s'." % self._cname)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def slot_status(self):
-        raise NotImplementedError("slot_status() not implemented for Scheduler '%s'." % self._cname)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _allocate_slot(self, cores_requested):
-        raise NotImplementedError("_allocate_slot() not implemented for Scheduler '%s'." % self._cname)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _release_slot(self, opaque_slots):
-        raise NotImplementedError("_release_slot() not implemented for Scheduler '%s'." % self._cname)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _try_allocation(self, cu):
-        """
-        Attempt to allocate cores for a specific CU.  If it succeeds, send the
-        CU off to the ExecutionWorker.
-        """
-
-        # needs to be locked as we try to acquire slots, but slots are freed
-        # in a different thread.  But we keep the lock duration short...
-        with self._slot_lock :
-
-            # schedule this unit, and receive an opaque handle that has meaning to
-            # the LRMS, Scheduler and LaunchMethod.
-            cu['opaque_slots'] = self._allocate_slot(cu['description']['cores'])
-
-        if not cu['opaque_slots']:
-            # signal the CU remains unhandled
-            return False
-
-        # got an allocation, go off and launch the process
-        self._prof.prof('schedule', msg="allocated", uid=cu['_id'])
-        self._log.info("slot status after allocated  : %s" % self.slot_status ())
-
-        # FIXME: if allocation succeeded, then the unit will likely advance to
-        #        executing soon.  Advance will do a blowup before puching -- but
-        #        that will also *drop* units.  We need to unschedule those.
-        #        self.unschedule(cu_dropped), and should probably do that right
-        #        here?  Not sure if this is worth a dropping-hook on component
-        #        level...
-        return True
-
-
-    # --------------------------------------------------------------------------
-    #
-    def reschedule_cb(self, topic, msg):
-        # we ignore any passed CU.  In principle the cu info could be used to
-        # determine which slots have been freed.  No need for that optimization
-        # right now.  This will become interesting once reschedule becomes too
-        # expensive.
-
-        cu = msg
-
-        self._prof.prof('reschedule', uid=self._pilot_id)
-        self._log.info("slot status before reschedule: %s" % self.slot_status())
-
-        # cycle through wait queue, and see if we get anything running now.  We
-        # cycle over a copy of the list, so that we can modify the list on the
-        # fly
-        for cu in self._wait_pool[:]:
-
-            if self._try_allocation(cu):
-
-                # allocated cu -- advance it
-                self.advance(cu, rp.EXECUTING_PENDING, publish=True, push=True)
-
-                # remove it from the wait queue
-                with self._wait_lock :
-                    self._wait_pool.remove(cu)
-                    self._prof.prof('unqueue', msg="re-allocation done", uid=cu['_id'])
-            else:
-                # Break out of this loop if we didn't manage to schedule a task
-                break
-
-        # Note: The extra space below is for visual alignment
-        self._log.info("slot status after  reschedule: %s" % self.slot_status ())
-        self._prof.prof('reschedule done')
-
-
-    # --------------------------------------------------------------------------
-    #
-    def unschedule_cb(self, topic, msg):
-        """
-        release (for whatever reason) all slots allocated to this CU
-        """
-
-        cu = msg
-        self._prof.prof('unschedule', uid=cu['_id'])
-
-        if not cu['opaque_slots']:
-            # Nothing to do -- how come?
-            self._log.warn("cannot unschedule: %s (no slots)" % cu)
-            return
-
-        self._log.info("slot status before unschedule: %s" % self.slot_status ())
-
-        # needs to be locked as we try to release slots, but slots are acquired
-        # in a different thread....
-        with self._slot_lock :
-            self._release_slot(cu['opaque_slots'])
-            self._prof.prof('unschedule', msg='released', uid=cu['_id'])
-
-        # notify the scheduling thread, ie. trigger a reschedule to utilize
-        # the freed slots
-        # FIXME: we don't have a reschedule pubsub, yet.  A local queue
-        #        should in principle suffice though.
-        self.publish('reschedule', cu)
-
-        # Note: The extra space below is for visual alignment
-        self._log.info("slot status after  unschedule: %s" % self.slot_status ())
-
-
-    # --------------------------------------------------------------------------
-    #
-    def clone_cb(self, unit, name=None, mode=None, prof=None, logger=None):
-
-        if mode == 'output':
-
-            # so, this is tricky: we want to clone the unit after scheduling,
-            # but at the same time don't want to have all clones end up on the
-            # same core -- so the clones should be scheduled to a different (set
-            # of) core(s).  But also, we don't really want to schedule, that is
-            # why we blow up on output, right?
-            #
-            # So we fake scheduling.  This assumes the 'self.slots' structure as
-            # used by the continuous scheduler, wo will likely only work for
-            # this one (FIXME): we walk our own index into the slot structure,
-            # and simply assign that core, be it busy or not.
-            #
-            # FIXME: This method makes no attempt to set 'task_slots', so will
-            # not work properly for some launch methods.
-            #
-            # This is awful.  I mean, really awful.  Like, nothing good can come
-            # out of this.  Ticket #902 should be implemented, it will solve
-            # this problem much cleaner...
-
-            if prof: prof.prof      ('clone_cb', uid=unit['_id'])
-            else   : self._prof.prof('clone_cb', uid=unit['_id'])
-
-            slot = self.slots[self._clone_slot_idx]
-
-            unit['opaque_slots']['task_slots'][0] = '%s:%d' \
-                    % (slot['node'], self._clone_core_idx)
-          # self._log.debug(' === clone cb out : %s', unit['opaque_slots'])
-
-            if (self._clone_core_idx +  1) < self._lrms_cores_per_node:
-                self._clone_core_idx += 1
-            else:
-                self._clone_core_idx  = 0
-                self._clone_slot_idx += 1
-
-                if self._clone_slot_idx >= len(self.slots):
-                    self._clone_slot_idx = 0
-
-
-    # --------------------------------------------------------------------------
-    #
-    def drop_cb(self, unit, name=None, mode=None, prof=None, logger=None):
-
-        if mode == 'output':
-            # we only unscheduler *after* scheduling.  Duh!
-
-            if prof:
-                prof.prof('drop_cb', uid=unit['_id'])
-            else:
-                self._prof.prof('drop_cb', uid=unit['_id'])
-
-            self.unschedule_cb(topic=None, msg=unit)
-
-
-
-    # --------------------------------------------------------------------------
-    #
-    def work(self, cu):
-
-      # self.advance(cu, rp.AGENT_SCHEDULING, publish=True, push=False)
-        self.advance(cu, rp.ALLOCATING      , publish=True, push=False)
-
-        # we got a new unit to schedule.  Either we can place it
-        # straight away and move it to execution, or we have to
-        # put it on the wait queue.
-        if self._try_allocation(cu):
-            self._prof.prof('schedule', msg="allocation succeeded", uid=cu['_id'])
-            self.advance(cu, rp.EXECUTING_PENDING, publish=True, push=True)
-
-        else:
-            # No resources available, put in wait queue
-            self._prof.prof('schedule', msg="allocation failed", uid=cu['_id'])
-            with self._wait_lock :
-                self._wait_pool.append(cu)
-
-
-
-# ==============================================================================
-#
-class SchedulerContinuous(AgentSchedulingComponent):
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        self.slots = None
-
-        AgentSchedulingComponent.__init__(self, cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        if not self._lrms_node_list:
-            raise RuntimeError("LRMS %s didn't _configure node_list." % self._lrms.name)
-
-        if not self._lrms_cores_per_node:
-            raise RuntimeError("LRMS %s didn't _configure cores_per_node." % self._lrms.name)
-
-        # Slots represents the internal process management structure.
-        # The structure is as follows:
-        # [
-        #    {'node': 'node1', 'cores': [p_1, p_2, p_3, ... , p_cores_per_node]},
-        #    {'node': 'node2', 'cores': [p_1, p_2, p_3. ... , p_cores_per_node]
-        # ]
-        #
-        # We put it in a list because we care about (and make use of) the order.
-        #
-        self.slots = []
-        for node in self._lrms_node_list:
-            self.slots.append({
-                'node': node,
-                # TODO: Maybe use the real core numbers in the case of
-                # non-exclusive host reservations?
-                'cores': [FREE for _ in range(0, self._lrms_cores_per_node)]
-            })
-
-
-    # --------------------------------------------------------------------------
-    #
-    def slot_status(self):
-        """Returns a multi-line string corresponding to slot status.
-        """
-
-        slot_matrix = ""
-        for slot in self.slots:
-            slot_matrix += "|"
-            for core in slot['cores']:
-                if core == FREE:
-                    slot_matrix += "-"
-                else:
-                    slot_matrix += "+"
-        slot_matrix += "|"
-        return {'timestamp' : rpu.timestamp(),
-                'slotstate' : slot_matrix}
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _allocate_slot(self, cores_requested):
-
-        # TODO: single_node should be enforced for e.g. non-message passing
-        #       tasks, but we don't have that info here.
-        if cores_requested <= self._lrms_cores_per_node:
-            single_node = True
-        else:
-            single_node = False
-
-        # Given that we are the continuous scheduler, this is fixed.
-        # TODO: Argument can be removed altogether?
-        continuous = True
-
-        # Switch between searching for continuous or scattered slots
-        # Switch between searching for single or multi-node
-        if single_node:
-            if continuous:
-                task_slots = self._find_slots_single_cont(cores_requested)
-            else:
-                raise NotImplementedError('No scattered single node scheduler implemented yet.')
-        else:
-            if continuous:
-                task_slots = self._find_slots_multi_cont(cores_requested)
-            else:
-                raise NotImplementedError('No scattered multi node scheduler implemented yet.')
-
-        if not task_slots:
-            # allocation failed
-            return {}
-
-        self._change_slot_states(task_slots, BUSY)
-        task_offsets = self.slots2offset(task_slots)
-
-        return {'task_slots'   : task_slots,
-                'task_offsets' : task_offsets,
-                'lm_info'      : self._lrms_lm_info}
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Convert a set of slots into an index into the global slots list
-    #
-    def slots2offset(self, task_slots):
-        # TODO: This assumes all hosts have the same number of cores
-
-        first_slot = task_slots[0]
-        # Get the host and the core part
-        [first_slot_host, first_slot_core] = first_slot.split(':')
-        # Find the entry in the the all_slots list based on the host
-        slot_entry = (slot for slot in self.slots if slot["node"] == first_slot_host).next()
-        # Transform it into an index in to the all_slots list
-        all_slots_slot_index = self.slots.index(slot_entry)
-
-        return all_slots_slot_index * self._lrms_cores_per_node + int(first_slot_core)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _release_slot(self, opaque_slots):
-
-        if not 'task_slots' in opaque_slots:
-            raise RuntimeError('insufficient information to release slots via %s: %s' \
-                    % (self.name, opaque_slots))
-
-        self._change_slot_states(opaque_slots['task_slots'], FREE)
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Find a needle (continuous sub-list) in a haystack (list)
-    #
-    def _find_sublist(self, haystack, needle):
-        n = len(needle)
-        # Find all matches (returns list of False and True for every position)
-        hits = [(needle == haystack[i:i+n]) for i in xrange(len(haystack)-n+1)]
-        try:
-            # Grab the first occurrence
-            index = hits.index(True)
-        except ValueError:
-            index = None
-
-        return index
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Transform the number of cores into a continuous list of "status"es,
-    # and use that to find a sub-list.
-    #
-    def _find_cores_cont(self, slot_cores, cores_requested, status):
-        return self._find_sublist(slot_cores, [status for _ in range(cores_requested)])
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Find an available continuous slot within node boundaries.
-    #
-    def _find_slots_single_cont(self, cores_requested):
-
-        for slot in self.slots:
-            slot_node = slot['node']
-            slot_cores = slot['cores']
-
-            slot_cores_offset = self._find_cores_cont(slot_cores, cores_requested, FREE)
-
-            if slot_cores_offset is not None:
-              # self._log.info('Node %s satisfies %d cores at offset %d',
-              #               slot_node, cores_requested, slot_cores_offset)
-                return ['%s:%d' % (slot_node, core) for core in
-                        range(slot_cores_offset, slot_cores_offset + cores_requested)]
-
-        return None
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Find an available continuous slot across node boundaries.
-    #
-    def _find_slots_multi_cont(self, cores_requested):
-
-        # Convenience aliases
-        cores_per_node = self._lrms_cores_per_node
-        all_slots = self.slots
-
-        # Glue all slot core lists together
-        all_slot_cores = [core for node in [node['cores'] for node in all_slots] for core in node]
-        # self._log.debug("all_slot_cores: %s", all_slot_cores)
-
-        # Find the start of the first available region
-        all_slots_first_core_offset = self._find_cores_cont(all_slot_cores, cores_requested, FREE)
-        self._log.debug("all_slots_first_core_offset: %s", all_slots_first_core_offset)
-        if all_slots_first_core_offset is None:
-            return None
-
-        # Determine the first slot in the slot list
-        first_slot_index = all_slots_first_core_offset / cores_per_node
-        self._log.debug("first_slot_index: %s", first_slot_index)
-        # And the core offset within that node
-        first_slot_core_offset = all_slots_first_core_offset % cores_per_node
-        self._log.debug("first_slot_core_offset: %s", first_slot_core_offset)
-
-        # Note: We subtract one here, because counting starts at zero;
-        #       Imagine a zero offset and a count of 1, the only core used
-        #       would be core 0.
-        #       TODO: Verify this claim :-)
-        all_slots_last_core_offset = (first_slot_index * cores_per_node) +\
-                                     first_slot_core_offset + cores_requested - 1
-        self._log.debug("all_slots_last_core_offset: %s", all_slots_last_core_offset)
-        last_slot_index = (all_slots_last_core_offset) / cores_per_node
-        self._log.debug("last_slot_index: %s", last_slot_index)
-        last_slot_core_offset = all_slots_last_core_offset % cores_per_node
-        self._log.debug("last_slot_core_offset: %s", last_slot_core_offset)
-
-        # Convenience aliases
-        last_slot = self.slots[last_slot_index]
-        self._log.debug("last_slot: %s", last_slot)
-        last_node = last_slot['node']
-        self._log.debug("last_node: %s", last_node)
-        first_slot = self.slots[first_slot_index]
-        self._log.debug("first_slot: %s", first_slot)
-        first_node = first_slot['node']
-        self._log.debug("first_node: %s", first_node)
-
-        # Collect all node:core slots here
-        task_slots = []
-
-        # Add cores from first slot for this unit
-        # As this is a multi-node search, we can safely assume that we go
-        # from the offset all the way to the last core.
-        task_slots.extend(['%s:%d' % (first_node, core) for core in
-                           range(first_slot_core_offset, cores_per_node)])
-
-        # Add all cores from "middle" slots
-        for slot_index in range(first_slot_index+1, last_slot_index):
-            slot_node = all_slots[slot_index]['node']
-            task_slots.extend(['%s:%d' % (slot_node, core) for core in range(0, cores_per_node)])
-
-        # Add the cores of the last slot
-        task_slots.extend(['%s:%d' % (last_node, core) for core in range(0, last_slot_core_offset+1)])
-
-        return task_slots
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Change the reserved state of slots (FREE or BUSY)
-    #
-    def _change_slot_states(self, task_slots, new_state):
-
-        # Convenience alias
-        all_slots = self.slots
-
-        # logger.debug("change_slot_states: unit slots: %s", task_slots)
-
-        for slot in task_slots:
-            # logger.debug("change_slot_states: slot content: %s", slot)
-            # Get the node and the core part
-            [slot_node, slot_core] = slot.split(':')
-            # Find the entry in the the all_slots list
-            slot_entry = (slot for slot in all_slots if slot["node"] == slot_node).next()
-            # Change the state of the slot
-            slot_entry['cores'][int(slot_core)] = new_state
-
-
-
-# ==============================================================================
-#
-class SchedulerScattered(AgentSchedulingComponent):
-    # FIXME: implement
-    pass
-
-
-# ==============================================================================
-#
-class SchedulerTorus(AgentSchedulingComponent):
-
-    # TODO: Ultimately all BG/Q specifics should move out of the scheduler
-
-    # --------------------------------------------------------------------------
-    #
-    # Offsets into block structure
-    #
-    TORUS_BLOCK_INDEX  = 0
-    TORUS_BLOCK_COOR   = 1
-    TORUS_BLOCK_NAME   = 2
-    TORUS_BLOCK_STATUS = 3
-
-
-    # --------------------------------------------------------------------------
-    def __init__(self, cfg):
-
-        self.slots            = None
-        self._cores_per_node  = None
-
-        AgentSchedulingComponent.__init__(self, cfg)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-        if not self._lrms_cores_per_node:
-            raise RuntimeError("LRMS %s didn't _configure cores_per_node." % self._lrms.name)
-
-        self._cores_per_node = self._lrms_cores_per_node
-
-        # TODO: get rid of field below
-        self.slots = 'bogus'
-
-
-    # --------------------------------------------------------------------------
-    #
-    def slot_status(self):
-        """Returns a multi-line string corresponding to slot status.
-        """
-
-        slot_matrix = ""
-        for slot in self._lrms.torus_block:
-            slot_matrix += "|"
-            if slot[self.TORUS_BLOCK_STATUS] == FREE:
-                slot_matrix += "-" * self._lrms_cores_per_node
-            else:
-                slot_matrix += "+" * self._lrms_cores_per_node
-        slot_matrix += "|"
-        return {'timestamp': rpu.timestamp(),
-                'slotstate': slot_matrix}
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Allocate a number of cores
-    #
-    # Currently only implements full-node allocation, so core count must
-    # be a multiple of cores_per_node.
-    #
-    def _allocate_slot(self, cores_requested):
-
-        block = self._lrms.torus_block
-        sub_block_shape_table = self._lrms.shape_table
-
-        self._log.info("Trying to allocate %d core(s).", cores_requested)
-
-        if cores_requested % self._lrms_cores_per_node:
-            num_cores = int(math.ceil(cores_requested / float(self._lrms_cores_per_node))) \
-                        * self._lrms_cores_per_node
-            self._log.error('Core not multiple of %d, increasing to %d!',
-                           self._lrms_cores_per_node, num_cores)
-
-        num_nodes = cores_requested / self._lrms_cores_per_node
-
-        offset = self._alloc_sub_block(block, num_nodes)
-
-        if offset is None:
-            self._log.warning('No allocation made.')
-            return
-
-        # TODO: return something else than corner location? Corner index?
-        sub_block_shape     = sub_block_shape_table[num_nodes]
-        sub_block_shape_str = self._lrms.shape2str(sub_block_shape)
-        corner              = block[offset][self.TORUS_BLOCK_COOR]
-        corner_offset       = self.corner2offset(self._lrms.torus_block, corner)
-        corner_node         = self._lrms.torus_block[corner_offset][self.TORUS_BLOCK_NAME]
-
-        end = self.get_last_node(corner, sub_block_shape)
-        self._log.debug('Allocating sub-block of %d node(s) with dimensions %s'
-                       ' at offset %d with corner %s and end %s.',
-                        num_nodes, sub_block_shape_str, offset,
-                        self._lrms.loc2str(corner), self._lrms.loc2str(end))
-
-        return {'cores_per_node'      : self._lrms_cores_per_node,
-                'loadl_bg_block'      : self._lrms.loadl_bg_block,
-                'sub_block_shape_str' : sub_block_shape_str,
-                'corner_node'         : corner_node,
-                'lm_info'             : self._lrms_lm_info}
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Allocate a sub-block within a block
-    # Currently only works with offset that are exactly the sub-block size
-    #
-    def _alloc_sub_block(self, block, num_nodes):
-
-        offset = 0
-        # Iterate through all nodes with offset a multiple of the sub-block size
-        while True:
-
-            # Verify the assumption (needs to be an assert?)
-            if offset % num_nodes != 0:
-                msg = 'Sub-block needs to start at correct offset!'
-                self._log.exception(msg)
-                raise ValueError(msg)
-                # TODO: If we want to workaround this, the coordinates need to overflow
-
-            not_free = False
-            # Check if all nodes from offset till offset+size are FREE
-            for peek in range(num_nodes):
-                try:
-                    if block[offset+peek][self.TORUS_BLOCK_STATUS] == BUSY:
-                        # Once we find the first BUSY node we can discard this attempt
-                        not_free = True
-                        break
-                except IndexError:
-                    self._log.exception('Block out of bound. Num_nodes: %d, offset: %d, peek: %d.',
-                            num_nodes, offset, peek)
-
-            if not_free == True:
-                # No success at this offset
-                self._log.info("No free nodes found at this offset: %d.", offset)
-
-                # If we weren't the last attempt, then increase the offset and iterate again.
-                if offset + num_nodes < self._block2num_nodes(block):
-                    offset += num_nodes
-                    continue
-                else:
-                    return
-
-            else:
-                # At this stage we have found a free spot!
-
-                self._log.info("Free nodes found at this offset: %d.", offset)
-
-                # Then mark the nodes busy
-                for peek in range(num_nodes):
-                    block[offset+peek][self.TORUS_BLOCK_STATUS] = BUSY
-
-                return offset
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Return the number of nodes in a block
-    #
-    def _block2num_nodes(self, block):
-        return len(block)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _release_slot(self, (corner, shape)):
-        self._free_cores(self._lrms.torus_block, corner, shape)
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Free up an allocation
-    #
-    def _free_cores(self, block, corner, shape):
-
-        # Number of nodes to free
-        num_nodes = self._shape2num_nodes(shape)
-
-        # Location of where to start freeing
-        offset = self.corner2offset(block, corner)
-
-        self._log.info("Freeing %d nodes starting at %d.", num_nodes, offset)
-
-        for peek in range(num_nodes):
-            assert block[offset+peek][self.TORUS_BLOCK_STATUS] == BUSY, \
-                'Block %d not Free!' % block[offset+peek]
-            block[offset+peek][self.TORUS_BLOCK_STATUS] = FREE
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Follow coordinates to get the last node
-    #
-    def get_last_node(self, origin, shape):
-        ret = {}
-        for dim in self._lrms.torus_dimension_labels:
-            ret[dim] = origin[dim] + shape[dim] -1
-        return ret
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Return the number of nodes for the given block shape
-    #
-    def _shape2num_nodes(self, shape):
-
-        nodes = 1
-        for dim in self._lrms.torus_dimension_labels:
-            nodes *= shape[dim]
-
-        return nodes
-
-
-    # --------------------------------------------------------------------------
-    #
-    # Return the offset into the node list from a corner
-    #
-    # TODO: Can this be determined instead of searched?
-    #
-    def corner2offset(self, block, corner):
-        offset = 0
-
-        for e in block:
-            if corner == e[self.TORUS_BLOCK_COOR]:
-                return offset
-            offset += 1
-
-        return offset
-
-#===============================================================================
-#
-class SchedulerYarn(AgentSchedulingComponent):
-
-    # FIXME: clarify what can be overloaded by Scheduler classes
-
-    # --------------------------------------------------------------------------
-    #
-    def __init__(self, cfg):
-
-        AgentSchedulingComponent.__init__(self, cfg)
-
-    # --------------------------------------------------------------------------
-    #
-    def _configure(self):
-
-        #-----------------------------------------------------------------------
-        # Find out how many applications you can submit to YARN. And also keep
-        # this check happened to update it accordingly
-
-
-        #if 'rm_ip' not in self._cfg['lrms_info']:
-        #    raise RuntimeError('rm_ip not in lm_info for %s' \
-        #            % (self.name))
-
-        self._log.info('Checking rm_ip %s' % self._cfg['lrms_info']['lm_info']['rm_ip'])
-        self._rm_ip = self._cfg['lrms_info']['lm_info']['rm_ip']
-        self._service_url = self._cfg['lrms_info']['lm_info']['service_url']
-        self._rm_url = self._cfg['lrms_info']['lm_info']['rm_url']
-        self._client_node = self._cfg['lrms_info']['lm_info']['nodename']
-
-        sample_time = rpu.timestamp()
-        yarn_status = ul.urlopen('http://{0}:8088/ws/v1/cluster/scheduler'.format(self._rm_ip))
-
-        yarn_schedul_json = json.loads(yarn_status.read())
-
-        max_num_app = yarn_schedul_json['scheduler']['schedulerInfo']['queues']['queue'][0]['maxApplications']
-        num_app = yarn_schedul_json['scheduler']['schedulerInfo']['queues']['queue'][0]['numApplications']
-
-        #-----------------------------------------------------------------------
-        # Find out the cluster's resources
-        cluster_metrics = ul.urlopen('http://{0}:8088/ws/v1/cluster/metrics'.format(self._rm_ip))
-
-        metrics = json.loads(cluster_metrics.read())
-        self._mnum_of_cores = metrics['clusterMetrics']['totalVirtualCores']
-        self._mmem_size = metrics['clusterMetrics']['totalMB']
-        self._num_of_cores = metrics['clusterMetrics']['allocatedVirtualCores']
-        self._mem_size = metrics['clusterMetrics']['allocatedMB']
-
-        self.avail_app = {'apps':max_num_app - num_app,'timestamp':sample_time}
-        self.avail_cores = self._mnum_of_cores - self._num_of_cores
-        self.avail_mem = self._mmem_size - self._mem_size
-
-    # --------------------------------------------------------------------------
-    #
-    def slot_status(self):
-        """
-        Finds how many spots are left free in the YARN scheduler queue and also
-        updates if it is needed..
-        """
-        #-------------------------------------------------------------------------
-        # As it seems this part of the Scheduler is not according to the assumptions
-        # made about slot status. Keeping the code commented just in case it is
-        # needed later either as whole or art of it.
-        sample = rpu.timestamp()
-        yarn_status = ul.urlopen('http://{0}:8088/ws/v1/cluster/scheduler'.format(self._rm_ip))
-        yarn_schedul_json = json.loads(yarn_status.read())
-
-        max_num_app = yarn_schedul_json['scheduler']['schedulerInfo']['queues']['queue'][0]['maxApplications']
-        num_app = yarn_schedul_json['scheduler']['schedulerInfo']['queues']['queue'][0]['numApplications']
-        if (self.avail_app['timestamp'] - sample)>60 and \
-           (self.avail_app['apps'] != max_num_app - num_app):
-            self.avail_app['apps'] = max_num_app - num_app
-            self.avail_app['timestamp']=sample
-
-        return '{0} applications per user remaining. Free cores {1} Free Mem {2}'\
-        .format(self.avail_app['apps'],self.avail_cores,self.avail_mem)
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _allocate_slot(self, cores_requested,mem_requested):
-        """
-        In this implementation it checks if the number of cores and memory size
-        that exist in the YARN cluster are enough for an application to fit in it.
-        """
-
-        #-----------------------------------------------------------------------
-        # If the application requests resources that exist in the cluster, not
-        # necessarily free, then it returns true else it returns false
-        #TODO: Add provision for memory request
-        if (cores_requested+1) <= self.avail_cores and \
-              mem_requested<=self.avail_mem and \
-              self.avail_app['apps'] != 0:
-            self.avail_cores -=cores_requested
-            self.avail_mem -=mem_requested
-            self.avail_app['apps']-=1
-            return True
-        else:
-            return False
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _release_slot(self, opaque_slot):
-        #-----------------------------------------------------------------------
-        # One application has finished, increase the number of available slots.
-        #with self._slot_lock:
-        self._log.info('Releasing : {0} Cores, {1} RAM'.format(opaque_slot['task_slots'][0],opaque_slot['task_slots'][1]))
-        self.avail_cores +=opaque_slot['task_slots'][0]
-        self.avail_mem +=opaque_slot['task_slots'][1]
-        self.avail_app['apps']+=1
-        return True
-
-
-
-    # --------------------------------------------------------------------------
-    #
-    def _try_allocation(self, cu):
-        """
-        Attempt to allocate cores for a specific CU.  If it succeeds, send the
-        CU off to the ExecutionWorker.
-        """
-        #-----------------------------------------------------------------------
-        # Check if the YARN scheduler queue has space to accept new CUs.
-        # Check about racing conditions in the case that you allowed an
-        # application to start executing and before the statistics in yarn have
-        # refreshed, to send another one that does not fit.
-
-        # TODO: Allocation should be based on the minimum memor allocation per
-        # container. Each YARN application needs two containers, one for the
-        # Application Master and one for the Container that will run.
-
-        # needs to be locked as we try to acquire slots, but slots are freed
-        # in a different thread.  But we keep the lock duration short...
-        with self._slot_lock :
-
-            self._log.info(self.slot_status())
-            self._log.debug('YARN Service and RM URLs: {0} - {1}'.format(self._service_url,self._rm_url))
-
-            # We also need the minimum memory of the YARN cluster. This is because
-            # Java issues a JVM out of memory error when the YARN scheduler cannot
-            # accept. It needs to go either from the configuration file or find a
-            # way to take this value for the YARN scheduler config.
-
-            cu['opaque_slots']={'lm_info':{'service_url':self._service_url,
-                                            'rm_url':self._rm_url,
-                                            'nodename':self._client_node},
-                                'task_slots':[cu['description']['cores'],2048]
-                                            }
-
-            alloc = self._allocate_slot(cu['description']['cores'],2048)
-
-        if not alloc:
-            return False
-
-        # got an allocation, go off and launch the process
-        self._prof.prof('schedule', msg="allocated", uid=cu['_id'])
-        self._log.info("slot status after allocated  : %s" % self.slot_status ())
-
-        return True
-
-    # --------------------------------------------------------------------------
-    #
-    def work(self, cu):
-
-      # self.advance(cu, rp.AGENT_SCHEDULING, publish=True, push=False)
-        self._log.info("Overiding Parent's class method")
-        self.advance(cu, rp.ALLOCATING , publish=True, push=False)
-
-        # we got a new unit to schedule.  Either we can place it
-        # straight away and move it to execution, or we have to
-        # put it on the wait queue.
-        if self._try_allocation(cu):
-            self._prof.prof('schedule', msg="allocation succeeded", uid=cu['_id'])
-            self.advance(cu, rp.EXECUTING_PENDING, publish=False, push=True)
-
-        else:
-            # No resources available, put in wait queue
-            self._prof.prof('schedule', msg="allocation failed", uid=cu['_id'])
-            with self._wait_lock :
-                self._wait_pool.append(cu)
 
 
 
@@ -5328,7 +4313,7 @@ class AgentWorker(rpu.Worker):
         # a factory might be more appropriate (FIXME)
         cmap = {
             "AgentStagingInputComponent"  : AgentStagingInputComponent,
-            "AgentSchedulingComponent"    : AgentSchedulingComponent,
+            "AgentSchedulingComponent"    : rp.agent.Scheduler,
             "AgentExecutingComponent"     : AgentExecutingComponent,
             "AgentStagingOutputComponent" : AgentStagingOutputComponent
             }
