+++ src/radical/pilot/agent/scheduler/__init__.py
@@ -1,0 +1,7 @@
+
+__copyright__ = "Copyright 2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+from .base import AgentSchedulingComponent as Scheduler
+
