+++ src/radical/pilot/agent/scheduler/scattered.py
@@ -1,0 +1,24 @@
+
+__copyright__ = "Copyright 2013-2016, http://radical.rutgers.edu"
+__license__   = "MIT"
+
+
+import radical.utils as ru
+
+from ... import utils     as rpu
+from ... import states    as rps
+from ... import constants as rpc
+
+from .base import AgentSchedulingComponent, FREE, BUSY
+
+
+# ==============================================================================
+#
+class Scattered(AgentSchedulingComponent):
+    
+    def __init__(self):
+        raise NotImplementedError('Scattered agent scheduler is not yet implemented')
+
+
+# ------------------------------------------------------------------------------
+
